+++ src/features/mobility-graph/constants.js
@@ -1,0 +1,36 @@
+export const typeSelection = [
+    {
+      id: 'parks',
+      label: 'Parks'
+    },
+    {
+      id: 'residential',
+      label: 'Residential'
+    },
+    {
+      id: 'retail-and-recreation',
+      label: 'Retail and Recreation'
+    },
+    {
+      id: 'transit-stations',
+      label: 'Transit Stations'
+    },
+    {
+      id: 'workplaces',
+      label: 'Workplaces'
+    },
+    {
+      id: 'grocery-and-pharmacy',
+      label: 'Grocery and Pharmacy'
+    }
+];
+
+export const idMap = {
+    'Retail and Recreation': 'retail-and-recreation',
+    'Grocery and Pharmacy': 'grocery-and-pharmacy',
+    Parks: 'parks',
+    'Transit Stations': 'transit-stations',
+    Workplaces: 'workplaces',
+    Residential: 'residential'
+};
+
